Moved SFINAE input iterator check to separated file
